Run nightly
Skip-checkpatch: true
Skip-build-leap15-rpm: true
Skip-nlt: true
Skip-coverity-test: true
Skip-test-centos-rpms: true
Skip-build-leap15-gcc: true
Skip-build-leap15-icc: true
Skip-build-ubuntu-gcc: true
Skip-build-ubuntu-clang: true
Skip-build-centos7-gcc-debug: true
Skip-build-centos7-gcc-release: true
Skip-build-centos7-clang: true
Skip-build-leap15-clang: true
Skip-build-ubuntu20-rpm: true
Skip-ubuntu20-rpm: true
Skip-func-test-leap15: true
Skip-unit-tests: true
Test-tag-vm: daily_regression,-hw
Test-tag-hw-small: daily_regression,hw,small
Test-tag-hw-medium: daily_regression,hw,medium,ib2
Test-tag-hw-large: daily_regression,hw,large

+++ utils/run_test.sh
@@ -197,7 +197,10 @@ if [ -d "/mnt/daos" ]; then
     if [ -f "/tmp/test.cov" ]; then
         rm /tmp/test.cov
     fi
-    cp "${DAOS_BASE}"/test.cov /tmp/
+
+    if [ -f "${DAOS_BASE}/test.cov" ]; then
+        cp "${DAOS_BASE}"/test.cov /tmp/
+    fi
 
     # Reporting
     if [ $failed -eq 0 ]; then
